Youplay/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2019.3.9
chardet==3.0.4
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.8
kombu==3.0.37
numpy==1.16.3
pandas==0.23.4
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.3
xlrd==1.0.0
